Delete in Changeset

+++ src/core/database/common.repository.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { ID, Session } from '../../common';
+import { node, relation } from 'cypher-query-builder';
+import { ID, isIdLike, ServerException, Session } from '../../common';
 import { DatabaseService } from './database.service';
 
 /**
@@ -13,7 +14,26 @@ export class CommonRepository {
     return await this.db.checkDeletePermission(id, session);
   }
 
-  async deleteNode(objectOrId: { id: ID } | ID) {
-    await this.db.deleteNode(objectOrId);
+  async deleteNode(objectOrId: { id: ID } | ID, changeset?: ID) {
+    if (!changeset) {
+      await this.db.deleteNode(objectOrId);
+      return;
+    }
+    try {
+      const id = isIdLike(objectOrId) ? objectOrId : objectOrId.id;
+      await this.db
+        .query()
+        .match([node('node', 'BaseNode', { id })])
+        .match([node('changeset', 'Changeset', { id: changeset })])
+        .merge([
+          node('changeset'),
+          relation('out', 'rel', 'changeset'),
+          node('node'),
+        ])
+        .setValues({ 'rel.active': true, 'rel.deleting': true })
+        .run();
+    } catch (e) {
+      throw new ServerException('Failed to remove node in changeset', e);
+    }
   }
 }
